refactor code, add responses

+++ pkg/user/repo.go
@@ -4,20 +4,24 @@ import (
 	"omega/engine"
 )
 
+// Repo for injecting engine
 type Repo struct {
 	Engine engine.Engine
 }
 
+// ProvideRepo is used in wire
 func ProvideRepo(engine engine.Engine) Repo {
 	return Repo{Engine: engine}
 }
 
+// FindAll users
 func (p *Repo) FindAll() (users []User, err error) {
 	err = p.Engine.DB.Find(&users).Error
 	return
 }
 
-func (p *Repo) FindByID(id uint) (user User, err error) {
+// FindByID for user
+func (p *Repo) FindByID(id uint64) (user User, err error) {
 	err = p.Engine.DB.First(&user, id).Error
 
 	user.Extra = struct {
@@ -31,16 +35,19 @@ func (p *Repo) FindByID(id uint) (user User, err error) {
 	return
 }
 
+// FindByUsername for user
 func (p *Repo) FindByUsername(username string) (user User, err error) {
 	err = p.Engine.DB.Where("username = ?", username).First(&user).Error
 	return
 }
 
+// Save user
 func (p *Repo) Save(user User) (u User, err error) {
 	err = p.Engine.DB.Save(&user).Scan(&u).Error
 	return
 }
 
+// Delete user
 func (p *Repo) Delete(user User) (err error) {
 	err = p.Engine.DB.Delete(&user).Error
 	return
